core_modules/EHSEngine/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pika==1.3.2
influxdb-client==1.36.1
scikit-learn==1.3.0
tensorflow==2.13.0
pandas==2.1.0
numpy==1.24.3
pydantic==1.10.13
joblib==1.3.2
pyyaml==6.0.1
